review: create standalone iter_torch_batches file
