+++ python/ray/data/BUILD.bazel
@@ -804,6 +804,20 @@ py_test(
 )
 
 py_test(
+    name = "test_iter_torch_batches",
+    size = "small",
+    srcs = ["tests/test_iter_torch_batches.py"],
+    tags = [
+        "exclusive",
+        "team:data",
+    ],
+    deps = [
+        ":conftest",
+        "//:ray_lib",
+    ],
+)
+
+py_test(
     name = "test_iterator",
     size = "small",
     srcs = ["tests/test_iterator.py"],
